+++ src/QueryBuilder/AbstractDQLQueryBuilder.php
@@ -536,16 +536,16 @@ abstract class AbstractDQLQueryBuilder implements DQLQueryBuilderInterface
             Query::class => QueryExpressionBuilder::class,
             Param::class => ParamBuilder::class,
             Expression::class => ExpressionBuilder::class,
-            Condition\NotCondition::class => Condition\Builder\NotConditionBuilder::class,
+            Condition\NotCondition::class => Condition\NotConditionBuilder::class,
             Condition\Conjunction\AndCondition::class => Condition\Conjunction\AndConditionBuilder::class,
             Condition\Conjunction\OrCondition::class => Condition\Conjunction\OrConditionBuilder::class,
-            Condition\BetweenCondition::class => Condition\Builder\BetweenConditionBuilder::class,
-            Condition\InCondition::class => Condition\Builder\InConditionBuilder::class,
-            Condition\LikeCondition::class => Condition\Builder\LikeConditionBuilder::class,
-            Condition\ExistsCondition::class => Condition\Builder\ExistsConditionBuilder::class,
-            SimpleCondition::class => Condition\Builder\SimpleConditionBuilder::class,
-            HashCondition::class => Condition\Builder\HashConditionBuilder::class,
-            Condition\BetweenColumnsCondition::class => Condition\Builder\BetweenColumnsConditionBuilder::class,
+            Condition\BetweenCondition::class => Yiisoft\Db\QueryBuilder\Condition\BetweenConditionBuilder::class,
+            Condition\InCondition::class => Yiisoft\Db\QueryBuilder\Condition\InConditionBuilder::class,
+            Condition\LikeCondition::class => Yiisoft\Db\QueryBuilder\Condition\LikeConditionBuilder::class,
+            Condition\ExistsCondition::class => Yiisoft\Db\QueryBuilder\Condition\ExistsConditionBuilder::class,
+            SimpleCondition::class => Condition\SimpleConditionBuilder::class,
+            HashCondition::class => Yiisoft\Db\QueryBuilder\Condition\HashConditionBuilder::class,
+            Condition\BetweenColumnsCondition::class => Yiisoft\Db\QueryBuilder\Condition\BetweenColumnsConditionBuilder::class,
             JsonExpression::class => JsonExpressionBuilder::class,
             ArrayExpression::class => ArrayExpressionBuilder::class,
             StructuredExpression::class => StructuredExpressionBuilder::class,
