+++ tests/Db/QueryBuilder/Condition/Builder/BetweenColumnsConditionBuilderTest.php
@@ -6,7 +6,7 @@ namespace Yiisoft\Db\Tests\Db\QueryBuilder\Condition\Builder;
 
 use PHPUnit\Framework\TestCase;
 use Yiisoft\Db\QueryBuilder\Condition\BetweenColumnsCondition;
-use Yiisoft\Db\QueryBuilder\Condition\Builder\BetweenColumnsConditionBuilder;
+use Yiisoft\Db\QueryBuilder\Condition\BetweenColumnsConditionBuilder;
 use Yiisoft\Db\Tests\Support\TestTrait;
 
 /**
